+++ server/package.json
@@ -16,7 +16,7 @@
         "@types/mongodb": "^3.1.22",
         "@types/oauth": "^0.9.1",
         "ajv": "^4.11.2",
-        "apollo-server-express": "^2.16.1",
+        "apollo-server-express": "^2.9.12",
         "compression": "^1.6.2",
         "connect-flash": "^0.1.1",
         "cors": "^2.8.5",
@@ -25,7 +25,7 @@
         "express-graphql": "^0.7.1",
         "express-session": "^1.15.6",
         "git-rev-sync": "^1.8.0",
-        "graphql": "^14.7.0",
+        "graphql": "^14.5.8",
         "graphql-tools": "^6.0.14",
         "gulp-copy": "^4.0.1",
         "gulp-tslint": "^8.1.4",
@@ -44,7 +44,7 @@
         "request": "^2.88.0",
         "serve-static": "^1.11.1",
         "tslint": "^5.15.0",
-        "typescript": "^3.3.3333",
+        "typescript": "^3.7.2",
         "uuid": "^3.3.2",
         "webpack": "^4.29.6"
     },
