+++ server/src/directives/index.ts
@@ -1,0 +1,5 @@
+import AuthDirective from './auth';
+
+export default {
+    auth: AuthDirective
+}
