+++ server/src/models/index.ts
@@ -1,0 +1,6 @@
+
+export { default as User } from './user'
+
+export { default as Team } from './team'
+
+export { default as Not } from './notification'
