+++ server/src/resolvers/index.ts
@@ -1,0 +1,5 @@
+import user from './user';
+import team from './team';
+import source from './source'
+
+export default [source, user, team]
