+++ server/src/resolvers/team/index.ts
@@ -1,0 +1,14 @@
+import { updateTeam } from './mutations'
+import { getTeam, getTeams } from './queries';
+
+const resolvers = {
+    Query: {
+        teams: getTeams,
+        team: getTeam
+      },
+      Mutation: {
+        update_team: updateTeam
+      }
+}
+
+export default resolvers
