+++ server/src/resolvers/team/mutations/index.ts
@@ -1,0 +1,1 @@
+export { default as updateTeam } from './updateTeam'
