+++ server/src/resolvers/team/queries/index.ts
@@ -1,0 +1,3 @@
+export { default as getTeam } from './getTeam'
+
+export { default as getTeams } from './getTeams'
