+++ server/src/resolvers/user/mutations/index.ts
@@ -1,0 +1,1 @@
+export { default as updateUser } from './updateUser'
