+++ server/src/resolvers/user/queries/index.ts
@@ -1,0 +1,5 @@
+export { default as getUser } from './getUser'
+
+export { default as getUsers } from './getUsers'
+
+export { default as getUserProfile } from './getUserProfile'
