+++ server/src/typeDefs/index.ts
@@ -1,0 +1,6 @@
+import user from './user'
+import team from './team'
+import notification from './notification'
+import root from './root'
+
+export default [root, user, team, notification]
