+++ server/src/types/index.ts
@@ -1,0 +1,7 @@
+export * from './notification'
+
+export * from './mongoose'
+
+export * from './team'
+
+export * from './user'
