+++ server/tsconfig.json
@@ -20,7 +20,8 @@
       "noUnusedParameters": false,
       "noImplicitReturns": true,
       "noFallthroughCasesInSwitch": true,
-      "esModuleInterop": true
+      "esModuleInterop": true,
+      "skipLibCheck": true
     },
     "include": [
       "src/**/*"
